Add awk and drawpile program names to bash variables
Use mawk explicitly because gawk is default awk in Ubuntu 20.04 but is not supported here.

Use python3 explicitly because it is supported and default in Ubuntu 20.04, while python2 is not installed and python command without any version does not exist. Also python2 is already past EOL, and new people are probably not using it anymore.